adding test for signup endpoint
additionally refactoring chatroom packages
